requirements.txt
Django==4.1
psycopg2_binary==2.9.3
djangorestframework==3.13.1
django_cors_headers==3.12.0
PyJWT==2.4.0
django_debug_toolbar==3.4.0
django_extensions==3.1.5
celery==5.2.6
requests==2.27.1
django_filter==21.1
redis==4.3.1
drf_yasg==1.20.0
factory_boy==3.2.1
boto3==1.24.17
sendgrid==6.9.7
stripe==5.0.0